+++ src/index.js
@@ -1,3 +1,3 @@
 export * from './router.js';
-export { readableStore as route } from './store.js';
+export { route } from './route.svelte.js';
 export * from './link.js';